+++ Mosqueta/mosqueta v 2.depend
@@ -1,0 +1,13 @@
+# depslib dependency file v1.0
+1526759854 source:c:\users\wilson\documents\codeblock\mosqueta v 2\main.cpp
+	<stdio.h>
+	<stdlib.h>
+	<string.h>
+	<time.h>
+
+1526781535 source:c:\users\warri\onedrive\documentos\github\mosqueta\mosqueta\main.cpp
+	<stdio.h>
+	<stdlib.h>
+	<string.h>
+	<time.h>
+
