+++ Mosqueta/mosqueta v 2.layout
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8" standalone="yes" ?>
+<CodeBlocks_layout_file>
+	<FileVersion major="1" minor="0" />
+	<ActiveTarget name="Debug" />
+	<File name="main.cpp" open="1" top="1" tabpos="1" split="0" active="1" splitpos="0" zoom_1="0" zoom_2="0">
+		<Cursor>
+			<Cursor1 position="10344" topLine="165" />
+		</Cursor>
+		<Folding>
+			<Collapse line="52" />
+			<Collapse line="77" />
+		</Folding>
+	</File>
+</CodeBlocks_layout_file>
